requirements.txt
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# LangChain / LangGraph
langchain==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
langgraph==0.1.5

# Embeddings & Vector Store
sentence-transformers==3.0.0
faiss-cpu==1.8.0

# Document parsing
pypdf==4.2.0
beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.32.2

# Observability
langfuse==2.36.1

# Utilities
httpx==0.27.0
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4
rich==13.7.1
typer==0.12.3

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
